Add an option to output to a file all links for Mastodon pages
This includes both the JavaScript based pages and the API requests
that the JavaScript based pages will be making.

Also include links to external sites referenced in the threads.

This can be useful to archive Mastodon pages to archive.org,
in a way that will work when viewing them on the Wayback Machine.