Add support for Mastodon off-site redirect pages
Mastodon links to off-site posts first HTTP redirect to a URL
that only links to the real off-site URL, for "safety".

For example [1] redirects to [2] which manual-redirects to [3].
Of these [1] is supported and neither of [2] or [3] are supported yet.
In a browser, clicking on [1] will leave you with not yet supported [2],
so it is useful to have support for [2] since the redirect there is manual
and [3] isn't supported yet.

1. https://kolektiva.social/@cacu@pixelfed.de/114240859464421461
2. https://kolektiva.social/redirect/statuses/114240859464421461
3. https://pixelfed.de/p/cacu/811263815440830587